requirements.txt
streamlit==1.41.1
pymupdf==1.24.14
pillow==10.4.0
openai==1.57.0
pydantic==2.9.2
reportlab==4.2.5
bcrypt==4.2.0
python-dotenv==1.0.1
